pocs/layered_reasoning_bot/requirements.txt
colorlog>=6.8.0
numpy>=1.26.3
pyyaml>=6.0.1
litellm>=1.15.0
tenacity>=8.2.3
matplotlib>=3.8.2
seaborn>=0.13.1
python-dotenv>=1.0.0 